requirements.txt
Mastodon.py>=1.0.5
